PERF-604: Surface test selection within a patch build in system_perf.yml

+++ bin/run-benchRun.sh
@@ -27,8 +27,13 @@ rm -f ../../reports.tgz
 
 source ${BINDIR}/setup-workloads.sh
 
+# Copy over the test_control.yml from repo if we don't already have one.
+if [ ! -e test_control.yml ]
+then
+    cp $DSI_PATH/test_control/test_control.benchRun.yml test_control.yml
+fi
+
 # PERF-531. Generating config file for mission control.
-cp $DSI_PATH/test_control/test_control.benchRun.yml test_control.yml
 python $BINDIR/config_test_control.py
 echo "Generated mc.json"
 scp -oStrictHostKeyChecking=no -i $PEMFILE  workloads.yml $SSHUSER@$mc:./workloads/
